Hawk: use id -> repo___path for the file index
The previous approach of using repo -> path had one problem:
it wouldn't let us do proper pattern matching on the repository
URL.

This commit adds a new constant to GraphModelBatchInjector with
the separator, which is used in most of the code dealing with
the file index. However, Neo4JDatabase#compareWithLocalFiles
also needs to access the file index and it can't depend on the
org.hawk.graph, so it also has its own copy of the constant.