Use PDE UI Ant tasks to build the update site
The Modelio plugins weren't working properly, because their classes depended
on an @objid annotation that was made available as an external library
(lib/javadesigner.jar) and not as part of the plugin classpath. Using "Build
All" from the site.xml would recompile everything using only plugin classpaths,
and that'd produce Modelio JARs with compile errors for the @objid annotations.

The @objid annotation use SOURCE retention, so it should be enough with simply
reusing the .class files produced in the workspace to produce the Hawk update
sites.

This commits replaces the site.xml file with an Ant buildfile with two tasks:
"build-site" that collects the features and plugins, and "publish-site" that
takes the collected features and plugins and produces a zipped p2 update site.
It is necessary to wait a bit after running build-site before running
publish-site, since pde.exportFeatures does not wait for the PDE Export to
complete.