Use new extension point to decouple HManager/HModel from LocalHawk
RemoteHawk needs to use the Thrift API, which is in the MONDO integration
project (as it includes the Thrift API of other components, and we will
also have a shared set of APIs for auth/user management).

To avoid a cyclic dependency between Hawk and the MONDO integration bits,
the best option is to have RemoteHawk be provided by a plugin within the
MONDO project.

This commit adds a new IHawkFactory interface + extension point to the
core plugin, which can be used from osgiserver to be able to provide new
IHawk implementations from other plugins. The core plugin has the default
implementation based on LocalHawk, so we'll always have at least one
option available.

The commit also extends the XStream-based configuration so it'll specify
the identifier of the IHawkFactory used. This is needed to reload the
configuration from scratch.

The commit has been tested to be working. The only remaining bit is to
provide a selector in the UI for the IHawkFactory to be used.