multiple additions and fixes, documented below:
added delete to igraphnodeindex so that derived/indexed attribute
indexes can be removed (for example if a metamodel is removed).
added exception catching in compositegraphchangelistener so errors in
the listeners do not affect hawk itself.
fixed error in synchronise where hawk would ignore the remaining
resources after failure.
added ability to write a program for derived attribute definition (used
to only allow expressions) -- assumes the program will use "return X;"
for defining the value of the attribute.
split some methods in deletionutils to utils as they do not delete
anything.
optimised dirtyderivedattributeslistner so that it only accesses
database indexes when it is asked for it, not on the fly.
added listeners when deleting model elements from metamodel removal
(only exposes element deletion not attr/ref).
fixed graphmodelbatchinjector bug when reverting back to a previous
version due to an error during batch insertion.