orientdb: fix two issues when accessing from UI
One problem was that the current OrientDB instance wasn't being activated
automatically when starting a transaction in some cases.

Another problem was that shutdown() did not close the underlying storage,
which caused problems with Windows file locks when using the Shift+Del
combination to forcefully delete a database in the Hawk view.