org.hawk.emfresource: fix Hawk+EMF-Splitter regression
ConcurrentModificationExceptions were being thrown when trying
to open a combobox for selecting the target object of a single-valued
reference in a Hawk resource. I fixed this by not adding newly
fetched objects that had a container in the graph to the root
contents of the resource.

However, this broke the EMF-Splitter integration, as objects that were
fetched directly by type would not be part of any Resource. Since Hawk
resources can be assumed to be read-only from the UI, I preferred to
have the integration working instead of the combobox, which ideally
should not have to iterate through the full resource in a dedicated UI.