modelio.exml: add 'hawkParent' and 'hawkChildren'
According to Antonin, Modelio compositions use containment on a per-case basis
for Associations and AssociationEnds, depending on their navigability. For this
reason, we can only rely on the PID of the MObjects, which represents a container
ref from EMF's point of view.

However, EMF generally treats container refs as derived from containment refs, so
we still need to provide a containment ref. An easy solution for this is to use the
recently added support for derived model references in Hawk: the graph stores the
OID->PID edge as "hawkParent" and then computes "hawkChildren" (the containment ref
EMF wants) from it.

This means that the EXML driver now needs to hook into Hawk and add "hawkChildren"
as a derived attribute of all Modelio classes.