modelio.exml: go back to enabling singleton objects - ensures model element uniqueness across fragments
Conversation between Antonio and Antonin (Softeam):

Antonio: I assume that UIDs are only really unique within a Modelio project and not across multiple projects: is that correct?

An UID is always unique but you can find 2 time the same UID in the repository ( in this case , it represent the same element).

- In Modelio an element is hosted by a Model Fragment and the same fragment can be deployed on several project.
- The same element (identified by is UID) can appear in several fragment ( for example in an SVN fragment and in a RAMC).
- In a project, an element (and it’s UID) can only  appear one time.

Antonio: Can I assume that if the URL of the .exml file is protocol://a/b/c/Project/data/.../*.exml, "Project" is the name of the project?

Yes but I think that is not what you need. A project is just an artificial construction  defined by a set of fragments.
It might be better to scope an element based on its fragment.