Add GPLv3 as secondary license
We first went with the default example with GPLv2 + Classpath exception, but
then realized we had quite a few Apache 2.0 libraries and we needed GPLv3
instead. Then we realized Neo4j did not come with the Classpath exception, so
we could not offer it here either - the secondary license is plain GPLv3.