GraphModelInserter: reload nodes that should have their derived features updated
If a node is changed in batch mode and requires its derived features to be updated,
later on it may fail during the update if there was a change between tx and batch
modes (e.g. in Neo4j). This was the last issue causing the 11059->11056 marker drop
in the EMFSplitter result20k dataset.