[REINDEX] Change file, ofType, ofKind edge labels to avoid clashes
While trying out the samples from @orichalque, I noticed that one of
the metamodels uses a 'file' EReference, which clashes with our own
'file' edges from the model element nodes to the file nodes.

This commit changes the names of the file, ofType, and ofKind internal
edges to add a '_hawk' prefix. We assume that metamodels will not have
attributes starting with this prefix.

Users will need to re-index after this change.