Use concurrent data structures for indexer components
I noticed that in some new tests that used multiple VCSes, the sync
on the first VCS would fail if you tried to add a VCS while that was
going on. Technically we would not allow that in the GUI, but we
should have something slightly more robust when deployed as a server.

As a fix, I have changed the data structures to use the concurrent
versions, which have consistent views upon iteration even if they
are modified in the middle of it.