core: create ITimeAwareGraphNodeIndex interface with travelInTime(long)
I experimented a bit with just extending the graph DB database to
create indices with a fixed time and have the default methods use
the graph DB's current time, but that broke quite a few tests.

Our code assumes that unless specified otherwise, IGraphNodeIndex
instances are reusable for various queries, and this includes reuse
across changes in the default timepoint of the backend.

The safest bet is to have a travelInTime(long) method for the indices,
which will produce a new instance that will use that as its default
timepoint and leave the old instance as-is. This way, we can get both
the old behaviour of relying on the backend time, and the new behaviour
of specifying a custom timepoint.