[REINDEX] Attribute derivation is marked with timestamp
GreyCat databases are smart in that they will not produce a new version
of a node if nothing changes. However, that creates a problem with derived
attributes, which are kept in a separate node from the element node. If only
the derived attribute node changes and the element node does not change, the
user will have no way to get at the derived attribute from the element node
in EOL.

To work around this, element nodes will be marked with a timestamp every
time a derived property is computed for them. This will ensure that
time-aware backends save a new version of the node as well.